Simplify the resolution of `scf.forall` created by split reductions. (#21422)
Earlier implementation of resolution of the `scf.forall` turned out to
be more involved than it needed to be. A simpler solution was to fold
the `scf.forall` into the workgroup mapping loop and increase the rank
of the loop appropriately. This PR implements that solution.

The current approach is kept as well, but disabled by default. It also
kicks in when there is no workgroup mapping for the split-reduction loop
to merge with, though this probably also needs to be fixed. The reason
for keeping the current approach is that folding the split-reduction
`scf.forall` and workgroup-mapped `scf.forall` effectively "sinks" the
computation with the split-reduction loop (but outside the workgroup
mapped loop) into the workgroup mapped loop. This might not always be
desirable, so the previous approach is kept as is for now.

Signed-off-by: MaheshRavishankar <mravisha@amd.com>