[Codegen][Encoding] Fix generic op materialization with 0D tensors (#21545)
The generic op materialization for data tiling will fail if any of the
operands resolve to identity layout, when the output operand does not,
but 0D tensors will always resolve to identity layout in data tiling.
This PR adds a special case for 0D tensors to fix the bug.

Fixes https://github.com/iree-org/iree/issues/21540

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>