Reapply "[Codegen] Fix dominance issues blocking consumer fusions (#21551)" (#21637)
Reapplies
https://github.com/iree-org/iree/commit/2688a7971c8f27eed803d0a5a8272d8240074b38
with a bug fix for https://github.com/iree-org/iree/issues/21622.
Operations were being moved by the rewriter before the
FuseTilableConsumer pattern was guaranteed to succeed, which added the
ops to the pattern worklist every time the pattern failed. This caused
an infinite recursion of the pattern application. The fix is to move the
ops without the rewriter, so they will not be added to the worklist.
This op movement can also be removed once the upstream fusion
transformation is refactored to properly handle multi-result scf.forall
loops.

---------

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>