[Codegen][GPU] Use arithmetic intensity to guide gemm size categorization - Step 1 (#21638)
This is the first PR to implement
https://github.com/iree-org/iree/discussions/21506 strategy 1 Subgroup
favor for seed selection.

This PR adds two optional fields to the chip attribute:
- peakMemoryBandwidthTBs: Single float point to indicated the bandwidth
in unit of tb/s
- peakPerfTFLOPs: Dictionary with key being the data type and value
being the tflops/s

With the added data in MI100, MI200, MI300 and Navi4 arch, next PR can
use the tflops and tbps to categorize the GEMMs into three buckets
(small, medium and large) and set seeds which are winners from
collection of 478 convolutions.

---------

Signed-off-by: jerryyin <zhuoryin@amd.com>