[CPU] Improve TileRootAndFuseProducerConsumer pass and deprecate TileAndFuse pass.  (#21674)
- Add an option to `TileAndFuseProducerConsumer` to start with last
available op. The current requirement is that the anchor op can only be
consumers or root op itself. It needs to be relaxed for producers, but
we match the behavior as LLVMCPUTileAndFuse for now.
- Rename the pass as `TileAndFuseProducerConsumer`.
- Switch all the uses of `TileAndFuse` pass to
`TileAndFuseProducerConsumer` pass.
- Adapt `tile_and_fuse.mlir` tests for the new option.
- Update scalable vector tests.
- Deprecate `LLVMCPUTileAndFuse` pass.
- The distribution field is added to
`tile_and_fuse_producer_consumer_anchoring_root_op.mlir` because the
pass uses it to determine which op is the root op.

The deprecation of `LLVMCPUTileAndFuse` pass is included in the revision
because the `tile_and_fuse.mlir` tests are adapted. Otherwise, we'd
either have dup tests or less test converage.

---------

Signed-off-by: hanhanW <hanhan0912@gmail.com>