[Codegen] Add matmul and batched matmul to list of ops to generalize (#21720)
Changes in this PR are 

1) carry 'compilation info' across from linalg.matmul (or other named
op) to linalg.generic when running the generalize pass. See for example
the test `lowering_matmul_promotion.mlir`

2) Refactoring the SPIRV configuration selection. This now fails if a
matmul/batch_matmul appears in specialized form.

3) One nvidia test now uses a different pipeline. It matches earlier in
the selection process for something more specialized. Looking at the
history of the config selected, it looks like it used to choose a more
specialised config anyway:
https://github.com/iree-org/iree/blame/6fcec315ac5f93ab7a240f57aa17772e8cc90f87/compiler/src/iree/compiler/Codegen/SPIRV/test/config_nvidia_matmul.mlir#L61

---------

Signed-off-by: James Newling <james.newling@gmail.com>