[Codegen][IGEMM] Fix and preserve padding dim order for convs (#21772)
Fix for the first case in https://github.com/iree-org/iree/issues/21660.

The previous logic was trying to determine the position for each
convolution dimensions and assign padding values accordingly which is
fragile given the fact there could be different kinds of convolutions in
generic forms. This PR adds a mapping from dims of a convolution to the
corresponding dims in the GEMM space which is more robust for figuring
the dims for padding on convolutions.

---------

Signed-off-by: yzhang93 <zhyuhang88@gmail.com>
Signed-off-by: Max Dawkins <max.dawkins@gmail.com>
Co-authored-by: Max Dawkins <max.dawkins@gmail.com>