[Codegen] Rewrite test so LLVMGPUWarpReduction is not used (#21770)
There is some logic (introduced in
https://github.com/iree-org/iree/pull/20310) that relies on finding a `
iree_tensor_ext.dispatch.tensor.store` op to trigger the
LLVMGPUVectorDistribute configuration to kick in. If no such 'beacon' op
is found, we currently fall through to LLVMGPUWarpReduction.

This PR just rejiggles the test so that the IR is the expected state
(the state which it will be in when the full pipeline is run) so that
that we're on the right pass to use LLVMGPUVectorDistribute. I can
follow this PR up with a refactoring of the logic to make it more robust

Signed-off-by: James Newling <james.newling@gmail.com>